+++ inst/examples/ex-image.R
@@ -1,0 +1,3 @@
+## RGB
+ternary_plot(NULL, xlab = "Red", ylab = "Green", zlab = "Blue")
+ternary_image(f = rgb, n = 20, palette = FALSE)